refactor pr comments
