+++ app/src/organisms/Devices/ModuleCard/MagneticModuleData.tsx
@@ -1,11 +1,16 @@
 import * as React from 'react'
 import { useTranslation } from 'react-i18next'
-import { C_BLUE, C_SKY_BLUE, Text } from '@opentrons/components'
+import {
+  C_BLUE,
+  C_SKY_BLUE,
+  FONT_SIZE_CAPTION,
+  Text,
+} from '@opentrons/components'
 import {
   MAGNETIC_MODULE_V1,
   MAGNETIC_MODULE_V2,
 } from '@opentrons/shared-data/js/constants'
-import { StatusLabel } from './StatusLabel'
+import { StatusLabel } from '../../../atoms/StatusLabel'
 
 interface MagModuleProps {
   moduleStatus: string
@@ -23,18 +28,19 @@ export const MagneticModuleData = (
     //    TODO Immediately: iconColor should be animated
     <>
       <StatusLabel
-        moduleStatus={moduleStatus}
+        status={moduleStatus}
         backgroundColor={C_SKY_BLUE}
-        iconColor={moduleStatus === 'disengaged' ? C_BLUE : C_BLUE}
+        iconColor={C_BLUE}
       />
-      <Text fontSize="0.625rem">
-        {moduleModel === MAGNETIC_MODULE_V2
-          ? t('magdeck_gen2_height', {
-              height: moduleHeight,
-            })
-          : t('magdeck_gen1_height', {
-              height: moduleHeight,
-            })}
+      <Text fontSize={FONT_SIZE_CAPTION}>
+        {t(
+          moduleModel === MAGNETIC_MODULE_V2
+            ? 'magdeck_gen2_height'
+            : 'magdeck_gen1_height',
+          {
+            height: moduleHeight,
+          }
+        )}
       </Text>
     </>
   )
