+++ app/src/organisms/Devices/ModuleCard/__tests__/MagneticModuleData.test.tsx
@@ -3,11 +3,11 @@ import * as React from 'react'
 import { renderWithProviders } from '@opentrons/components'
 
 import { i18n } from '../../../../i18n'
-import { StatusLabel } from '../StatusLabel'
+import { StatusLabel } from '../../../../atoms/StatusLabel'
 import { MagneticModuleData } from '../MagneticModuleData'
 import { mockMagneticModule } from '../../../../redux/modules/__fixtures__'
 
-jest.mock('../StatusLabel')
+jest.mock('../../../../atoms/StatusLabel')
 
 const mockStatusLabel = StatusLabel as jest.MockedFunction<typeof StatusLabel>
